Started working on logic for uploading video and redirecting to results

+++ Default.aspx
@@ -1,9 +1,12 @@
 ﻿<%@ Page Title="SNIPS | Home" Language="C#" MasterPageFile="MasterContent.master" %>
 
+<%@ Import Namespace="System.IO" %>
+
 <script runat="server">
 
     protected void Page_Load(object sender, System.EventArgs e) {
 
+        bool testingFFMpeg = true;
         if (!IsPostBack)
         {
 
@@ -11,15 +14,54 @@
         else
         {
             String map_path = HttpContext.Current.Server.MapPath(".");
-            if ((System.IO.File.Exists(map_path + "\\cut.mp4")))
-            {
-                System.IO.File.Delete(map_path + "\\cut.mp4");
-            }
-            String result = ProcessVideo.StartProcess("C:\\ffmpeg.exe", "-ss 00:09:00 -i " + map_path + "\\sample2.mp4 -to 00:03:00 -c copy " + map_path + "\\cut.mp4");
-            String result1 = ProcessVideo.StartProcess("C:\\ffmpeg.exe", "-ss 00:08:20 -i " + map_path + "\\sample2.mp4 -to 00:01:00 -c copy " + map_path + "\\cut1.mp4");  
-            //String results2 = ProcessVideo.StartProcess("C:\\ffprobe.exe", "-v error -show_entries format=size -of default=noprint_wrappers=1 "+ map_path +"\\cut.mp4");
 
-            //pResults.InnerText = results2;
+            if (testingFFMpeg)
+            {
+                if ((System.IO.File.Exists(map_path + "\\cut.mp4")))
+                {
+                    System.IO.File.Delete(map_path + "\\cut.mp4");
+                }
+                String result = ProcessVideo.StartProcess("C:\\ffmpeg.exe", "-ss 00:09:00 -i " + map_path + "\\sample2.mp4 -to 00:03:00 -c copy " + map_path + "\\cut.mp4");
+                String result1 = ProcessVideo.StartProcess("C:\\ffmpeg.exe", "-ss 00:08:20 -i " + map_path + "\\sample2.mp4 -to 00:01:00 -c copy " + map_path + "\\cut1.mp4");
+                //String results2 = ProcessVideo.StartProcess("C:\\ffprobe.exe", "-v error -show_entries format=size -of default=noprint_wrappers=1 "+ map_path +"\\cut.mp4");
+                //pResults.InnerText = results2;
+            }
+            else
+            {
+                if (files.PostedFile != null && files.PostedFile.ContentLength > 0)
+                {
+                    String guid = Guid.NewGuid().ToString();
+                    HttpContext.Current.Session["guid"] = guid;
+                    HttpContext.Current.Session["timestamps"] = hdnTimestamps.Value;
+                    HttpContext.Current.Session["windows"] = hdnWindows.Value;
+
+                    string file_name = System.IO.Path.GetFileName(files.PostedFile.FileName);
+                    String save_location = map_path + "\\videos\\" + guid;
+                    Directory.CreateDirectory(save_location);
+                    try
+                    {
+                        files.PostedFile.SaveAs(save_location + "\\" + file_name);
+                        String[] time_stamps = hdnTimestamps.Value.Split(',');
+                        String[] window_sizes = hdnWindows.Value.Split(',');
+                        for (int i = 0; i < time_stamps.Length; i++)
+                        {
+                            TimeSpan start = TimeSpan.Parse("00:" + time_stamps[i]);
+                            TimeSpan length = TimeSpan.Parse("00:00:" + window_sizes[i]);
+                            TimeSpan finish = start.Add(length);
+                            start = start.Subtract(length);
+                            ProcessVideo.StartProcess("C:\\ffmpeg.exe", "-ss " + start.ToString() + " -i " + save_location + "\\" + file_name + " -to " + finish.ToString() + " -c copy " + save_location + "\\Snippet_" + (i + 1) + ".mp4");
+                        }
+                        HttpContext.Current.Session["snipsNum"] = time_stamps.Length;
+                        Response.Redirect("/Results.aspx");
+                    }
+                    catch (Exception ex)
+                    {
+
+                    }
+
+
+                }
+            }
         }
     }
 </script>
@@ -96,8 +138,21 @@
             });
         });
     });
-    function SaveResponses() {
+    function saveValues() {
+        var time_stamps = "";
+        $('.times').each(function () {
+            time_stamps += $(this).val() + ",";
+        });
+        time_stamps = time_stamps.substr(0, time_stamps.length - 1);
 
+        var windows_sizes = "";
+        $('.windowsizes').each(function () {
+            windows_sizes += $(this).val() + ",";
+        });
+        windows_sizes = windows_sizes.substr(0, windows_sizes.length - 1);
+
+        $('#hdnTimestamps').val(time_stamps);
+        $('#hdnWindows').val(windows_sizes);
     }
 
     // create additional time inputs
@@ -132,6 +187,7 @@
     var myVideos = [];
     window.URL = window.URL || window.webkitURL;
     function setFileInfo(files) {
+        myVideos.pop();
         myVideos.push(files[0]);
         var video = document.createElement('video');
         video.preload = 'metadata';
@@ -193,7 +249,7 @@
 
         <div class="upload">
             <%--<output id="list"></output>--%>
-            <input type="file" id="files" name="files[]" accept=".mov,.mp4,.m4v" onchange="setFileInfo(this.files)">
+            <input type="file" id="files" name="files[]" accept=".mov,.mp4,.m4v" onchange="setFileInfo(this.files)" runat="server" />
             <label class="btn btn-default btn-lg" for="files">
                 <i class="fa fa-cloud-upload"></i>&nbsp Upload
             </label>
@@ -223,7 +279,7 @@
         </div>
 
         <div class="submit">
-            <input type ="button" class="btn btn-default btn-lg" onclick ="$('#frm').submit();" value="Submit" />
+            <input type ="button" class="btn btn-default btn-lg" onclick ="javascript: saveValues(); $('#frm').submit();" value="Submit" />
         </div>
 
         <button type="submit">Test Snippetting</button>
